EOS-18924: Send Pmsg to DTM0 services with persistent log.
The patch introduces sending of PERSISTENT messages to the other
participants that have persistent log. Previously (the happy-path
milestone), they were sent only to the client (originator).

Also, the patch addresses several related issues:
- It fixes initialisation of log (mkfs).
- It resolves some TODOs and addresses code style-related issues
  for the original code.

Historical perspective: the original version of this patch was
created as (#660), and later on it was updated as a part of
the work on the start-stop milestone (drlink integration).

Signed-off-by: Mehul Joshi <mehul.joshi@seagate.com>
Signed-off-by: Ivan Alekhin <ivan.alekhin@seagate.com>