EOS-21935: Hotfix for AST loop caused by dix_rop_one.
The patch is a short term solution for a problem
described in EOS-21935. The problem is related to the
situation where dix_rop_one_completed is pushed
into the forkq twice from different ticks.
It causes a look in the forkq, and then it leads
to a panic.

The patch addresses this problem by bypassing the forkq
and directly calling this function from the clink callback.
It is safe because the clink is triggered from another AST,
and this AST has the same sm group as DIX req and DTX.