client: fix failed devs calc on node(s) failure(s)
On nodes/pods failures the number of failed devices may
be calculated wrongly at motr/io_req.c:device_check().
The reason is that we check for the node issue only in
case of ECANCELED error from the target (but there may
other errors too).

Solution: check for the node state always also if it's
not ECANCELED error.

Signed-off-by: Andriy Tkachuk <andriy.tkachuk@seagate.com>