CORTX-30872 dix: add support of transient failures for GET
Problem:
Before merging of degraded DIX PUT support for transient
failures (PR https://github.com/Seagate/cortx-motr/pull/1545)
the degraded GET worked by a mistake: sending of requests
to offline targets + resend logic.
The PR mentioned above prevents sending of requests to
offline devices, so the offline unit is skipped. Other alive
targets are not selected due to dix_online_unit_choose()
function (limits the number of units to be selected to 1
to send a single request for DIX GET, allows to iterate
over parity units in case of server-side failures).
So if the first unit is offline then whole DIX operation
wails with -EIO immediately and no resend logic is triggered.
So need a proper way to handle transient failures for DIX GET.

Solution:
Mark a unit as unavailable (previously named "failed") if
corresponding device is OFFLINE, so that it is considered
by the dix_online_unit_choose() function.

Signed-off-by: Sergey Shilov <sergey.shilov@seagate.com>